Moved parse_utils to its own library
