+++ source/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_subdirectory(common)
 add_subdirectory(errors)
 add_subdirectory(lexer)
+add_subdirectory(parse)
 add_subdirectory(ast)
 add_subdirectory(sema)
 add_subdirectory(exec)
