+++ source/ast/parse_errors_observer.hpp
@@ -4,24 +4,18 @@
 
 #include "errors/error.hpp"
 #include "errors/errors_observer.hpp"
+#include "parse/parse_errors_observer.hpp"
 
 namespace cmsl::ast {
-class parse_errors_observer
+class parse_errors_observer : public parse::parse_errors_observer
 {
 public:
   virtual ~parse_errors_observer() = default;
 
-  virtual void raise_unexpected_end_of_file(lexer::token token) = 0;
-  virtual void raise_expected_token(lexer::token token,
-                                    lexer::token_type type) = 0;
-  virtual void raise_unexpected_token(lexer::token token) = 0;
-  virtual void raise_expected_keyword(lexer::token token,
-                                      lexer::token_type keyword) = 0;
   virtual void raise_expected_type(lexer::token token) = 0;
   virtual void raise_expected_parameter_declaration(lexer::token token) = 0;
   virtual void raise_expected_parameter(lexer::token token) = 0;
   virtual void raise_expected_block(lexer::token token) = 0;
-  virtual void raise_expected_expression(lexer::token token) = 0;
   virtual void raise_expected_designated_initializer(lexer::token token) = 0;
   virtual void raise_expected_unary_operator(lexer::token token) = 0;
   virtual void raise_expected_enumerator(lexer::token token) = 0;
